+++ src/examples/subscribe-multiple-live-bars-concurrent.example.ts
@@ -1,0 +1,29 @@
+import { merge, tap } from 'rxjs';
+import { cTraderX } from '../classes/client';
+import { ProtoOATrendbarPeriod } from '../classes/managers/symbols/proto/models/ProtoOATrendbarPeriod';
+
+// Subscribe live bars
+(async () => {
+    const client = new cTraderX();
+
+    await client.connect();
+    merge(
+        ...[
+            client.symbolsUpdates
+                .subscribeLiveTrendBars({
+                    period: ProtoOATrendbarPeriod.M1,
+                    // symbolId: 1, // EURUSD
+                    symbolId: 10026, // BTCUSD
+                })
+                .pipe(tap((event) => console.log(`Sub 1: ${event.period}`))),
+
+            client.symbolsUpdates
+                .subscribeLiveTrendBars({
+                    period: ProtoOATrendbarPeriod.M5,
+                    // symbolId: 1, // EURUSD
+                    symbolId: 10026, // BTCUSD
+                })
+                .pipe(tap((event) => console.log(`Sub 2: ${event.period}`))),
+        ],
+    ).subscribe();
+})();